Create ToDo component and move out of App
